Server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
firebase-admin==6.2.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
python-multipart==0.0.6
websockets==12.0
requests==2.31.0
